refactor(nbd-cow): fix logic, safety, and performance issues from review
Logic:
- fix check_bounds to use u64 length (was truncating to u32)
- remove dead code in write() (empty if body)
- merge cow_path into CowLayer::new() (was two-step init, easy to forget)

Safety:
- add MAX_REQUEST_LENGTH (32MB) guard in read/write handlers to prevent
  OOM from oversized requests
- use saturating_add in bounds check

Performance:
- change Arc<Mutex<CowLayer>> to Arc<RwLock<CowLayer>> so reads don't
  block each other across connections
- hold write lock through write+flush to avoid redundant concurrent flushes

Documentation:
- add note about sync file I/O in async context and when to consider
  spawn_blocking

+++ crates/nbd-cow/src/error.rs
@@ -11,7 +11,7 @@ pub enum NbdCowError {
     #[error("offset {offset} + length {length} exceeds device size {device_size}")]
     OutOfBounds {
         offset: u64,
-        length: u32,
+        length: u64,
         device_size: u64,
     },
